More refactor, Update CI
